extract crud to service
