+++ src/main/java/example/taskmanager/TaskController.java
@@ -1,0 +1,46 @@
+package example.taskmanager;
+
+import example.taskmanager.dto.TaskDTO;
+import example.taskmanager.service.TaskService;
+import org.springframework.http.ResponseEntity;
+import org.springframework.stereotype.Controller;
+import org.springframework.web.bind.annotation.*;
+
+import java.util.List;
+import java.util.Optional;
+
+@Controller
+@RequestMapping("/api/tasks")
+public class TaskController {
+    TaskService taskService;
+
+    public TaskController(TaskService taskService) {
+        this.taskService = taskService;
+    }
+
+    @GetMapping
+    public @ResponseBody List<TaskDTO> getTasks(@RequestParam(value = "completed", required = false) Optional<Boolean> completed) {
+        return taskService.getAllTasks(completed);
+    }
+
+    @GetMapping("/{id}")
+    public ResponseEntity<TaskDTO> getTask(@PathVariable("id") Long id) {
+        return taskService.getTaskById(id);
+    }
+
+    @PostMapping
+    public ResponseEntity<TaskDTO> postTask(@RequestBody TaskDTO taskDTO) {
+        TaskDTO createdTask = taskService.createTask(taskDTO);
+        return ResponseEntity.ok(createdTask);
+    }
+
+    @PutMapping("/{id}")
+    public ResponseEntity<TaskDTO> putTask(@PathVariable("id") Long id, @RequestBody TaskDTO taskDTO) {
+        return taskService.updateTask(id, taskDTO);
+    }
+
+    @DeleteMapping("/{id}")
+    public ResponseEntity<Void> deleteTask(@PathVariable("id") Long id) {
+        return taskService.removeTask(id);
+    }
+}
